Whatsnews-we can now see only the upcoming news

+++ components/WhatsNewViewAll/WhatsNewViewAll.jsx
@@ -8,14 +8,15 @@ const formatDate = (dateString) => {
   const day = date.getDate();
   const month = date.getMonth() + 1;
   const year = date.getFullYear();
-  return `${day}-${month}-${year}`;
+  return `${day}.${month}.${year}`;
 };
 
 const WhatsNewViewAll = () => {
   const [data, setData] = useState([]);
   const [isLoading, setLoading] = useState(true);
   const token = cookie.get("token");
-
+  var currentDate = new Date().toISOString().split('T')[0];
+  var formatCurrentDate = formatDate(currentDate);
   const handleDelete = async (id) => {
     debugger;
     try {
@@ -27,9 +28,12 @@ const WhatsNewViewAll = () => {
           token: token,
         },
       };
-      await axios.delete(`/api/whatsNew/${id}`, config);
-      const response = await axios.get("/api/whatsNew", config);
-      setData(response.data);
+      var result = confirm("Want to delete?");
+      if (result) {
+        await axios.delete(`/api/whatsNew/${id}`, config);
+        const response = await axios.get("/api/whatsNew", config);
+        setData(response.data);
+      }
     } catch (error) {
       console.error(error);
     }
@@ -50,7 +54,7 @@ const WhatsNewViewAll = () => {
 
         const sortedData = response.data
           .slice()
-          .sort((a, b) => new Date(b.date) - new Date(a.date));
+          .sort((b, a) => new Date(b.date) - new Date(a.date));
 
         setData(sortedData);
         setLoading(false);
@@ -62,44 +66,47 @@ const WhatsNewViewAll = () => {
 
     fetchData();
   }, [token]);
+  console.log(data);
+  console.log("currentDate " + currentDate + "\n");
+  // .toISOString().split('T')[0])
+  // data.map((whatsNew) => (console.log(Date.parse(whatsNew.date.split('T')[0]) > Date.parse(currentDate)))
 
   return (
     <div>
-      {data.map((whatsNew) => (
-        <div
-          key={whatsNew._id}
-          className="items-center w-full border-collapse bg-white"
-        >
-          <div className="flex m-2 py-4">
-            <div className="flex flex-[1] flex-col gap-2 p-2">
-              <div className="text-sm font-semibold">
-                {formatDate(whatsNew.date)}
+      {data.map((whatsNew) => (Date.parse(whatsNew.date.split('T')[0]) < Date.parse(currentDate)) ? (console.log(whatsNew))
+        :
+        (
+          <div
+            key={whatsNew._id}
+            className="items-center w-full border-collapse bg-white"
+          >
+            <div className="flex m-2 py-4">
+              <div className="flex flex-[1] flex-col gap-2 p-2">
+                <div className="text-sm font-semibold">
+                  {formatDate(whatsNew.date)}
+                </div>
+                <div className="text-xs font-light">
+                  <div>posted by</div>
+                  <div>{whatsNew.postedBy}</div>
+                </div>
               </div>
-              <div className="text-xs font-light">
-                <div>posted by</div>
-                <div>{whatsNew.postedBy}</div>
+              <div className="flex flex-[3] flex-col gap-2 p-2">
+                <div className="text-sm font-semibold">{whatsNew.title}</div>
+                <div className="text-xs font-light">{whatsNew.paragraph}</div>
               </div>
-            </div>
-            <div className="flex flex-[3] flex-col gap-2 p-2">
-              <div className="text-sm font-semibold">{whatsNew.title}</div>
-              <div className="text-xs font-light">{whatsNew.paragraph}</div>
-            </div>
-            <div className="flex flex-[1] items-end p-2">
-              <div className="flex h-fit text-sm font-semibold underline cursor-pointer">
-                Read More
-              </div>
-              <div>
-                <button
-                  onClick={() => handleDelete(whatsNew._id)}
-                  className="ml-2 self-start"
-                >
-                  <DeleteIcon />
-                </button>
+              <div className="flex flex-[1] items-end p-2">
+                <div>
+                  <button
+                    onClick={() => handleDelete(whatsNew._id)}
+                    className="ml-2 self-start"
+                  >
+                    <DeleteIcon />
+                  </button>
+                </div>
               </div>
             </div>
           </div>
-        </div>
-      ))}
+        ))}
     </div>
   );
 };
